refactor: 필요없는 api 메서드 정리
